+++ analysis/webservice/algorithms/doms/insitu.py
@@ -7,6 +7,18 @@ from datetime import datetime
 from webservice.algorithms.doms import config as insitu_endpoints
 
 
+def query_insitu_schema():
+    """
+    Query the "cdms_schema" insitu endpoint. This will return the JSON
+    schema used to construct the data, which will contain useful
+    metadata
+    """
+    schema_endpoint = insitu_endpoints.getSchemaEndpoint()
+    response = requests.get(schema_endpoint)
+    response.raise_for_status()
+    return response.json()
+
+
 def query_insitu(dataset, variable, start_time, end_time, bbox, platform, depth_min, depth_max,
                items_per_page=1000, session=None):
     """
@@ -44,13 +56,14 @@ def query_insitu(dataset, variable, start_time, end_time, bbox, platform, depth_
     insitu_response = {}
 
     # Page through all insitu results
-    next_page_url = insitu_endpoints.getEndpoint()
+    next_page_url = insitu_endpoints.getEndpoint(provider)
     while next_page_url is not None and next_page_url != 'NA':
-        logging.debug(f'Insitu request {next_page_url}')
         if session is not None:
             response = session.get(next_page_url, params=params)
         else:
             response = requests.get(next_page_url, params=params)
+
+        logging.debug(f'Insitu request {response.url}')
 
         response.raise_for_status()
         insitu_page_response = response.json()