+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SDAP-393: Included `insitu` in ingress based on the value of `insituAPI.enabled` in `values.yaml`
 - SDAP-371: Renamed `/domssubset` endpoint to `/cdmssubset`
 - SDAP-390: Updated NetCDF reader tool for data matchup and added user functionality.
+- SDAP-396: Added saildrone insitu api to matchup
+=======
 ### Changed
 -SDAP-390: Changed `/doms` to `/cdms` and `doms_reader.py` to `cdms_reader.py`
 ### Deprecated
@@ -21,4 +23,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix failing test_matchup unit test
 - Fixed null value for count in matchup response
 - SDAP-371: Fixed DOMS subset endpoint
+- SDAP-397: Added depth to matchup insitu output
+- SDAP-397: Pull ID from insitu api response for matchup
+- SDAP-397: Added null unit field for later use in matchup
+- Fixed issue where satellite satellite matchups failed
+- Fixed issue where null results were returned if more than "resultSizeLimit" matches are found
+- Preserve return value of `query_insitu_schema()` in matchup to avoid excessive API hits 
+- Fixed issue where satellite to satellite matchups with the same dataset don't return the expected result
+- Fixed CSV and NetCDF matchup output bug
+- Fixed NetCDF output switching latitude and longitude
 ### Security